requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#annotated-types==0.7.0    
#anyio==4.4.0    
azure-functions==1.20.0   
#certifi==2024.8.30
#distro==1.9.0    
#exceptiongroup==1.2.2    
groq==0.11.0   
#h11==0.14.0   
#httpcore==1.0.5    
#httpx==0.27.2   
#idna==3.8
pip==20.0.2      
setuptools==44.0.0   
#sniffio==1.3.1    
typing-extensions==4.12.2
#langchain   
#langchain-community

aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
black==24.10.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.8
langchain-core==0.3.23
langchain-openai==0.2.12
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.2
langfuse==2.55.0
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
#openai==0.28
openai==1.57.1
orjson==3.10.12
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
azure-cosmos
fastapi
uvicorn
gradio